move components into their own files and import them for use
